Bins file to pitch correct for C Major
This can be used in lieu of the bins code in 'LiveProcessing.m' to autotune for Concert C Major songs. If the reference pitch is changed, then the key will change too. In the example, we use 55 Hz as A, so using a reference frequency of B would result in the key of D instead of C.